chore: moves things to own file and renames
